+++ packages/jest-jasmine2/src/index.ts
@@ -121,9 +121,7 @@ async function jasmine2(
     .requireInternalModule<typeof import('./jestExpect')>(
       path.resolve(__dirname, './jestExpect.js'),
     )
-    .default({
-      expand: globalConfig.expand,
-    });
+    .default({expand: globalConfig.expand});
 
   if (globalConfig.errorOnDeprecated) {
     installErrorOnPrivate(environment.global);
@@ -222,7 +220,6 @@ const addSnapshotData = (
   return results;
 };
 
-// eslint-disable-next-line no-redeclare
 namespace jasmine2 {
   export type Jasmine = JestJasmine;
 }