+++ packages/jest-snapshot/src/types.ts
@@ -5,6 +5,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+/* eslint-disable local/ban-types-eventually */
+
 import type {MatcherState} from 'expect';
 import type SnapshotState from './State';
 
@@ -29,3 +31,6 @@ export type ExpectationResult = {
   pass: boolean;
   message: () => string;
 };
+
+export type BabelTraverse = typeof import('@babel/traverse').default;
+export type Prettier = typeof import('prettier');