+++ packages/jest-validate/src/exampleConfig.ts
@@ -12,8 +12,7 @@ const config: ValidationOptions = {
   condition: () => true,
   deprecate: () => false,
   deprecatedConfig: {
-    // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-    key: () => {},
+    key: (): string => 'Deprecation message',
   },
   error: () => {},
   exampleConfig: {key: 'value', test: 'case'},