+++ packages/jest-worker/src/Farm.ts
@@ -5,6 +5,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+/* eslint-disable local/ban-types-eventually */
+
 import {
   CHILD_MESSAGE_CALL,
   ChildMessage,